adobe-stock-integration#1806: refactoring and cover by mftf test
