+++ app/code/Magento/Catalog/Test/Mftf/Section/AdminCategoryContentSection.xml
@@ -14,6 +14,7 @@
         <element name="selectFromGalleryButton" type="button" selector="//*[@class='file-uploader-area']/label[text()='Select from Gallery']"/>
         <element name="uploadImageFile" type="input" selector=".file-uploader-area>input"/>
         <element name="imageFileName" type="text" selector=".file-uploader-filename"/>
+        <element name="imageFileMeta" type="text" selector=".file-uploader-meta"/>
         <element name="removeImageButton" type="button" selector=".file-uploader-summary .action-remove"/>
         <element name="AddCMSBlock" type="select" selector="//*[@name='landing_page']"/>
         <element name="description" type="input" selector="//*[@name='description']"/>
