--- app/code/Magento/MediaGalleryRenditions/etc/adminhtml/routes.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="admin">
-        <route id="media_gallery_renditions" frontName="media_gallery_renditions">
-            <module name="Magento_MediaGalleryRenditions" before="Magento_Backend" />
-        </route>
-    </router>
-</config>
